fix: update aggregate data types to Decimal for scientific notation support
Updated the coding keys for OHLCV data to use `Decimal` instead of `Int` or `Double`.

Previously, large values (such as volume) returned by the API in scientific notation (e.g., 3.73e+07) caused a decoding error: "Number is not representable in Swift." Switching to `Decimal` ensures high-precision financial data is parsed correctly and prevents crashes when encountering floating-point representations in the JSON payload.

+++ Sources/PolygonioSwift/Aggregates.swift
@@ -58,7 +58,7 @@ public struct AggregatesResponse : Decodable {
     }
     
     public struct Candle : Codable {
-        public var volume: Int
+        public var volume: Decimal
         public var VWAP: Decimal //VWAP (Volume Weighted Average Price)
         public var open: Decimal
         public var close: Decimal
